Extract parent class for indicators table
